LatLonData 클래스 추가 && geocoding.getLatLon() 메소드 구현

+++ src/main/java/com/quickorderservice/utiles/geo/GeoData.java
@@ -24,8 +24,8 @@ public class GeoData {
         private String roadAddress;
         private String jibunAddress;
         private String englishAddress;
-        private float x;
-        private float y;
+        private double x;
+        private double y;
         private float distance;
     }
